Fix bug where piece values were incorrect
The data types that were used to store the values of the pieces in the MoveData strcut were being overflowed, causing seemingly random values for pieces

+++ src/MoveGeneration.cpp
@@ -366,7 +366,7 @@ namespace MoveGeneration
         return nullptr;
     }
 
-    void getPieceData(Board* boardPtr, Bitboard** pieceBB, Byte* pieceValue, Byte square, Colour side)
+    void getPieceData(Board* boardPtr, Bitboard** pieceBB, int* pieceValue, Byte square, Colour side)
     {
         Bitboard squareBB = BB::boardSquares[square];
